feat: add a new controll to recenter map to user location
